refactor: DTO, BlackJackGame 제거
